Move `TBTC` interface to separate file
This way we achieve a clean `index.ts`
with all exports listed.
